requirements.txt
langchain
langchain-openai
langchain-community

openai
tiktoken

python-dotenv

pydantic

faiss-cpu

chromadb

unstructured

beautifulsoup4

requests

jupyter
ipykernel